Move BookForm component to the container folder
