+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import BookList from '../containers/BookList';
-import BookForm from './BookForm';
+import BookForm from '../containers/BookForm';
 
 const App = () => (
   <div>
